Improved data loading
Data now loads alpha numeric words and is smarter when loading sentences. It also maximizes data loading when making the vocab to maximize the data being loaded in. Instead of immediately stopping when the vocab is maximized, it only loads sentences with no new words and store the new sentences in a new file